+++ tests/Janus.Authentication.Tests/Identifiers/IdentifierServiceTests.cs
@@ -98,6 +98,7 @@ public sealed class IdentifierServiceTests : IAsyncDisposable
                 _mail,
                 _sms,
                 RestrictionKeySuppliers.None,
+                Considered.Nothing(_work, _clock),
                 new SmsBalance(_configuration, _sms, _balances, _work, _events, _clock),
                 _work,
                 _events,